new structure. Added new table, and enabled join operations on two created tables
